naksha-backend/requirements.txt
fastapi
python-multipart
uvicorn
pillow
ultralytics
numpy
